==> 1) process_data.ipynb <==
from os import listdir
import pandas as pd
from labeling.emotions import rage, joy, sadness, surprise, fear

months = {
    'stycznia': 1,
    'lutego': 2,
    'marca': 3,
    'kwietnia': 4,
    'maja': 5,
    'czerwca': 6,
    'lipca': 7,
    'sierpnia': 8,
    'września': 9,
    'października': 10,
    'listopada': 11,
    'grudnia': 12
}

def process_date(months, date):
    year = date.split('|')[0].split(' ')[2].strip()
    month = str(months[date.split('|')[0].split(' ')[1].strip()])
    day = date.split('|')[0].split(' ')[0].strip()
    
    try:
        hour = date.split('|')[1].strip()
        return pd.to_datetime(" ".join([year, month, day, hour]))
    except:
        return None
    
    

from re import search, sub
from unidecode import unidecode

# http://exp.lobi.nencki.gov.pl/nawl-analysis
more_emotion_keywords = pd.read_csv('labeling/nawl-analysis.csv')

# H - joy
# A - rage
# S - sadness
# D - surprise
# F - fear

Hapiness = more_emotion_keywords[more_emotion_keywords['category'] == 'H']['word'].values.tolist()
Anger = more_emotion_keywords[more_emotion_keywords['category'] == 'A']['word'].values.tolist()
Sadness = more_emotion_keywords[more_emotion_keywords['category'] == 'S']['word'].values.tolist()
Disgust = more_emotion_keywords[more_emotion_keywords['category'] == 'D']['word'].values.tolist()
Fear = more_emotion_keywords[more_emotion_keywords['category'] == 'F']['word'].values.tolist()

rage_regex = "|".join(rage + Anger)
joy_regex = "|".join(joy + Hapiness)
sadness_regex = "|".join(sadness + Sadness)
surprise_regex = "|".join(surprise + Disgust)
fear_regex = "|".join(fear + Fear)

def find_emotions(emotions, body):
    for i, emotion_regex in enumerate([rage_regex, joy_regex, sadness_regex, surprise_regex, fear_regex]):
        if search(emotion_regex, body):
            emotions[i] += 1
    return emotions

def label(comments):
    replies = 0

    upvotes = 0
    downvotes = 0

    emotions = [0,0,0,0,0]
    for comment in a['comments']:
        main = comment['main_comment']

        upvotes += int(main['upvotes'])
        downvotes += int(main['downvotes'])
        replies += 1

        emotions = find_emotions(emotions, sub(r"[^a-z| ]", "", unidecode(main['body']).lower()))
        for subcomment in comment['sub_comments']:
            try:
                upvotes += int(subcomment['upvotes'])
                downvotes += int(subcomment['downvotes'])
                replies += 1
            except:
                return None
            emotions = find_emotions(emotions, sub(r"[^a-z| ]", "", unidecode(main['body']).lower()))
    return [replies, upvotes, downvotes] + [emotion for emotion in emotions]

def remove_specials(body):
    return sub(r"[^a-z| ]", "", unidecode(body).lower())

with open('data/polish.stopwords.txt', 'r') as f:
    stop_words = f.read().split('\n')

def remove_stop_words(body):
    return [word for word in body.split(' ') if word not in stop_words and len(word) > 1]

def clean_words(body):
    return " ".join(remove_stop_words(remove_specials(body)))

all_articles =[]
for article in listdir('data/comments/'):
    with open('data/comments/'+article) as f:
        articles = []
        a = eval(f.read())['row']['comments']
        if a['comments'] != []:
        
            articles.append(article)
            articles.append(a['division'])
            articles.append(process_date(months, a['pub_date']))
            articles.append(a['author'].split(',')[0])

            articles.append(clean_words(a['title']))
            articles.append(clean_words(a['highlight']))
            articles.append(clean_words(a['content']))
            articles.append(clean_words(a['media_desc']))

            articles.append(a['media_type'])
            labels = label(a['comments'])
            if labels:
                for return_label in labels:
                    articles.append(return_label)
                all_articles.append(articles)

df = pd.DataFrame(all_articles, columns = [
    'id', 'div', 'date', 'author', 'title', 
    'highlight', 'content', 'media_desc', 'media_type',
    'replies', 'upvotes', 'downvotes', 
    'rage', 'joy', 'sadness', 'surprise', 'fear'
])
df['reactions'] = df['upvotes'] + df['downvotes']
df['reactions_sentiment'] = round(df['upvotes'] / df['reactions'], 2)

len(df)

df.head(3)

df.info()

def create_dict(unique):
    return { key: value for value, key in enumerate(unique) }

division_dict = create_dict(df['div'].unique())

author_dict = create_dict(df['author'].unique())

media_dict = create_dict(df['media_type'].unique())

from json import dump

with open('labeling/division_dict', 'w') as f:
    dump(division_dict, f)

# df['div'] = df['div'].map(division_dict)

with open('labeling/author_dict', 'w') as f:
    dump(author_dict, f)

# df['author'] = df['author'].map(author_dict)

with open('labeling/media_dict', 'w') as f:
    dump(media_dict, f)

# df['media_type'] = df['media_type'].map(media_dict)

df.head(3)

df.to_csv('labeling/fresh_data.csv', index = False)
