==> 2) lemmatize_articles.ipynb <==
import morfeusz2
import pandas as pd

m = morfeusz2.Morfeusz()

articles = pd.read_csv('labeling/fresh_data.csv')\
    [['id', 'title', 'highlight', 'content', 'media_desc']]\

articles.head(1)

articles.info()

from unidecode import unidecode

def lemmatize_column(column):
    lemmatized_column = [] 
    for record in column.values:
        lemmatized_row = []
        try:
            for word in record.split(' '):
                lemmatized_row.append(unidecode(m.analyse(word)[0][2][1].split(':')[0]).lower())
            lemmatized_column.append(" ".join(lemmatized_row))
        except:
            lemmatized_column.append(None)
        
    return pd.Series(lemmatized_column, name = column.name)

lem_df_list = []
for column in [u'content', u'highlight', u'media_desc', u'title']:
    lem_df_list.append(lemmatize_column(articles[column]))
lem_df_list.append(articles['id'])

lem_df = pd.concat(lem_df_list, axis = 1)

lem_df.head(5)

from collections import Counter

lem_df = lem_df.fillna('')

bag_of_records = lem_df['title'] + ' ' + lem_df['highlight'] + ' ' + lem_df['content'] + ' ' + lem_df['media_desc']
bag_of_words = [x for x in [row for row in bag_of_records] for x in x.split(' ')]

idf_dict = {word:freq for word, freq in dict(Counter(bag_of_words)).items() if freq > 2}

from json import dump

with open('labeling/idf_dict', 'w') as f:
    dump(idf_dict, f)

lem_df.to_csv('labeling/lemmatized_articles.csv', index = False)
