==> 3) filter_word2vec.ipynb <==
from json import load

with open('labeling/idf_dict') as f:
    idf = load(f)

vocabulary = list(idf.keys())

vocabulary[:6]

re_pattern = "|".join(vocabulary)

re_pattern[:30]

import pandas as pd

#file_iterator = pd.read_csv('~/polish_word2ve_300', iterator = True, chunksize = 10000)

filtered_word2vec = []
with open('../polish_word2vec_100', 'r') as f:
    for line in f:
        if line.split(' ')[0] in vocabulary:
            filtered_word2vec.append(line)

len(filtered_word2vec)

with open('../word2vec_test', 'w') as f:
    for item in filtered_word2vec:
        f.write("%s\n" % item)
