==> 4) apply_word2vec.ipynb <==
with open('/home/jupyter/word2vec_test', 'r') as f:
    raw_file = f.readlines()

word2vec = {}
for row in raw_file:
    word2vec[row.split(' ')[0]] = [float(x) for x in row.split(' ')[1:]]

vocabulary = list(word2vec.keys())

import pandas as pd
import numpy as np

df = pd.read_csv('labeling/lemmatized_articles.csv')

df.head(1)

df = df.fillna(' ')

errors = 0

def calculate_vec(column):
    series = []
    
    for record in column:
        n_words = 0
        vec = np.zeros(100, dtype = 'float32')
        if record != ' ':
            try:
                for word in record.split(' '):
                    if word in vocabulary:
                        vec = np.add(vec, word2vec[word])
                        n_words += 1
            except:
                errors += 1

        if n_words:
            vec = np.divide(vec, n_words)
        series.append(vec)
    
    return series

for column in ['content', 'highlight', 'media_desc', 'title']:
    pd.DataFrame(calculate_vec(df[column]), index = df['id']).to_csv('labeling/'+column+'_vec')

print(errors)
